refactor: add api/v1 to url
Move entities folder to src
